fixed tests a bit

+++ tests/testthat/test-ebirdgeo.R
@@ -5,7 +5,7 @@ test_that("ebirdgeo works correctly", {
   expect_is(egeo, "data.frame")
   expect_equal(ncol(egeo), 11)
   expect_is(egeo$comName, "character")
-  expect_is(egeo$howMany, "numeric")
+  expect_is(egeo$howMany, "integer")
   expect_is(ebirdgeo(lat = 42, lng = -76, max=10, provisional=TRUE, hotspot=TRUE), "data.frame")
   expect_equal(nrow(ebirdgeo(lat = 42, lng = -76, max=10, provisional=TRUE, hotspot=TRUE)), 10)
   expect_that(ebirdgeo(lat = 40, lng = -120, dist = 50), not(gives_warning()))
